07_django/99_json_response/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Django==3.2.18
djangorestframework==3.14.0
idna==3.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.2.0
pytz==2023.3
requests==2.28.2
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
